requirements.txt
torch>=2.0
numpy>=1.24
matplotlib>=3.7
pandas>=2.0
tqdm>=4.65
pyyaml>=6.0
scipy>=1.10
scikit-learn>=1.3
